requirements.txt
# requirements.txt
gradio
python-dotenv
pandas
Pillow
azure-cognitiveservices-vision-customvision
